+++ src/main/java/store/service/StoreService.java
@@ -1,15 +1,15 @@
-package store.domain;
+package store.service;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
+import store.domain.Stock;
 import store.domain.order.OrderItem;
 import store.domain.order.OrderItems;
 import store.domain.order.OrderStatus;
 
 public class StoreService {
-    // [콜라-10] 에서 이름과 수량을 캡처하는 패턴 정규식
     private static final String ORDER_ITEM_PATTERN_REGEX = "\\[(.+)-(\\d+)]";
 
     public OrderItems getOrderItems(String input) {
